add tp and hp percent detail move json save file to root reimport dbdiff to adapt to latest cn server version fixed expect val for additive action merged template exporting function

+++ Assets/Scrips/Elements/BattleUtil.cs
@@ -78,11 +78,11 @@ namespace Elements
 
     public static FloatWithEx FloatToIntReverseTruncate(FloatWithEx _num)
     {
-        return _num.Select(x => FloatToIntReverseTruncate(x));
+        return _num.Select(x => FloatToIntReverseTruncate((float)x));
     }
     public static FloatWithEx FloatToInt(FloatWithEx _num)
     {
-        return _num.Select(x => FloatToInt(x));
+        return _num.Select(x => FloatToInt((float)x));
     }
     public static eUnitRespawnPos SearchRespawnPos(
       eUnitRespawnPos _basePos,
